requirements.txt
swarms
openai
pdfminer.six